Extract FieldEditor from ViewEditor
