+++ packages/ui/react-ui-data/src/components/ViewEditor/ViewEditor.tsx
@@ -2,26 +2,17 @@
 // Copyright 2024 DXOS.org
 //
 
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 
-import { FormatEnum, FormatEnums, type MutableSchema, S } from '@dxos/echo-schema';
+import { type MutableSchema, S } from '@dxos/echo-schema';
 import { Button, Icon, useTranslation, type ThemedClassName } from '@dxos/react-ui';
 import { List } from '@dxos/react-ui-list';
 import { ghostHover, mx } from '@dxos/react-ui-theme';
-import {
-  createUniqueFieldForView,
-  getPropertySchemaForFormat,
-  type FieldProjection,
-  FieldSchema,
-  type FieldType,
-  type ViewType,
-  ViewProjection,
-  type PropertyType,
-} from '@dxos/schema';
+import { createUniqueFieldForView, FieldSchema, type FieldType, type ViewType, ViewProjection } from '@dxos/schema';
 import { arrayMove } from '@dxos/util';
 
 import { translationKey } from '../../translations';
-import { Form, FormInput } from '../Form';
+import { FieldEditor } from '../Field/FieldEditor';
 
 const grid = 'grid grid-cols-[32px_1fr_32px] min-bs-[2.5rem] rounded';
 
@@ -37,45 +28,16 @@ export type ViewEditorProps = ThemedClassName<{
 export const ViewEditor = ({ classNames, schema, view, readonly }: ViewEditorProps) => {
   const { t } = useTranslation(translationKey);
   const projection = useMemo(() => new ViewProjection(schema, view), [schema, view]);
-  const [field, setField] = useState<FieldType | undefined>();
-
-  const fieldProperties = useMemo<PropertyType>(() => {
-    if (!field) {
-      return;
-    }
-
-    const { props } = projection.getFieldProjection(field.property);
-    return props;
-  }, [field, view]);
-  const [{ fieldSchema }, setSchema] = useState({ fieldSchema: getPropertySchemaForFormat(fieldProperties?.format) });
-
-  // TODO(burdon): Update object type (field) when format changes (get from FormatSchema map).
-  // TODO(burdon): Handle changes to `property` (e.g., uniqueness)?
-  const handleValueChanged = useCallback((values: PropertyType) => {
-    setSchema({ fieldSchema: getPropertySchemaForFormat(values?.format) });
-  }, []);
-
-  useEffect(() => {
-    handleValueChanged(fieldProperties);
-  }, [fieldProperties]);
-
-  const handleAdditionalValidation = useCallback(
-    ({ property }: PropertyType) => {
-      if (property && view.fields.find((f) => f.property === property && f !== field)) {
-        return [{ path: 'property', message: `'${property}' is not unique.` }];
-      }
-    },
-    [view.fields, field],
-  );
+  const [selectedField, setSelectedField] = useState<FieldType>();
 
   const handleSelect = useCallback((field: FieldType) => {
-    setField((f) => (f === field ? undefined : field));
+    setSelectedField((f) => (f === field ? undefined : field));
   }, []);
 
   const handleAdd = useCallback(() => {
     const field = createUniqueFieldForView(view);
     view.fields.push(field);
-    setField(field);
+    setSelectedField(field);
   }, [view]);
 
   const handleMove = useCallback(
@@ -89,13 +51,8 @@ export const ViewEditor = ({ classNames, schema, view, readonly }: ViewEditorPro
     (field: FieldType) => {
       const idx = view.fields.findIndex((f) => field.property === f.property);
       view.fields.splice(idx, 1);
-      setField(undefined);
+      setSelectedField(undefined);
     },
-    [view.fields],
-  );
-
-  const handleSet = useCallback(
-    (field: FieldType, props: FieldProjection) => projection.setFieldProjection({ field, props }),
     [view.fields],
   );
 
@@ -127,37 +84,7 @@ export const ViewEditor = ({ classNames, schema, view, readonly }: ViewEditorPro
         )}
       </List.Root>
 
-      {!fieldSchema ? (
-        <div>Schema not implemented for {fieldProperties?.format ?? 'undefined'}</div>
-      ) : (
-        field &&
-        fieldProperties && (
-          <Form<PropertyType>
-            key={field.property}
-            autoFocus
-            values={fieldProperties}
-            schema={fieldSchema}
-            additionalValidation={handleAdditionalValidation}
-            onValuesChanged={handleValueChanged}
-            onSave={(props) => handleSet(field, props)}
-            Custom={(props) => (
-              <>
-                {/* TODO(burdon): Move property field here. */}
-                <FormInput<PropertyType>
-                  property='format'
-                  label={t('field format label')}
-                  placeholder={t('field format label')}
-                  options={FormatEnums.filter((value) => value !== FormatEnum.None).map((value) => ({
-                    value,
-                    label: t(`format ${value}`),
-                  }))}
-                  {...props}
-                />
-              </>
-            )}
-          />
-        )
-      )}
+      {selectedField && <FieldEditor field={selectedField} projection={projection} view={view} />}
 
       {/* TODO(burdon): Option. */}
       {!readonly && (
